Consolidate dashboard e2e specs (#43613)
* consolidate dashboard e2e specs

* fix import
