--- e2e/test/scenarios/dashboard/reproductions/29076-dashboard-card-drill-sandbox.cy.spec.js
@@ -1,54 +1,0 @@
-import { USER_GROUPS, SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { ORDERS_DASHBOARD_ID } from "e2e/support/cypress_sample_instance_data";
-const { ALL_USERS_GROUP, COLLECTION_GROUP } = USER_GROUPS;
-import {
-  describeEE,
-  restore,
-  visitDashboard,
-  setTokenFeatures,
-} from "e2e/support/helpers";
-
-const { PRODUCTS_ID, PRODUCTS } = SAMPLE_DATABASE;
-
-describeEE("issue 29076", () => {
-  beforeEach(() => {
-    restore("default-ee");
-
-    cy.intercept("/api/dashboard/*/dashcard/*/card/*/query").as("cardQuery");
-
-    cy.signInAsAdmin();
-    setTokenFeatures("all");
-    cy.updatePermissionsGraph({
-      [ALL_USERS_GROUP]: {
-        [SAMPLE_DB_ID]: {
-          "view-data": "blocked",
-          "create-queries": "no",
-        },
-      },
-      [COLLECTION_GROUP]: {
-        [SAMPLE_DB_ID]: {
-          "view-data": "unrestricted",
-          "create-queries": "query-builder",
-        },
-      },
-    });
-    cy.sandboxTable({
-      table_id: PRODUCTS_ID,
-      attribute_remappings: {
-        attr_uid: ["dimension", ["field", PRODUCTS.ID, null]],
-      },
-    });
-    cy.signInAsSandboxedUser();
-  });
-
-  it("should be able to drilldown to a saved question in a dashboard with sandboxing (metabase#29076)", () => {
-    visitDashboard(ORDERS_DASHBOARD_ID);
-    cy.wait("@cardQuery");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Orders").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Visualization").should("be.visible");
-  });
-});
